Adapt code to use Rusty-LevelDB
