+++ packages/core/crates/core-packet/src/interaction.rs
@@ -1084,7 +1084,7 @@ mod tests {
         sync::{Arc, Mutex},
         time::Duration,
     };
-    use utils_db::{db::DB, leveldb::rusty::RustyLevelDbShim};
+    use utils_db::{db::DB, rusty::RustyLevelDbShim};
     use utils_log::debug;
     use utils_types::{
         primitives::{Address, Balance, BalanceType, Snapshot, U256},
@@ -1153,35 +1153,35 @@ mod tests {
         )
     }
 
-    fn create_dbs(amount: usize) -> Vec<Arc<Mutex<rusty_leveldb::DB>>> {
+    fn create_dbs(amount: usize) -> Vec<RustyLevelDbShim> {
         (0..amount)
-            .map(|i| {
-                Arc::new(Mutex::new(
-                    rusty_leveldb::DB::open(format!("test_db_{i}"), rusty_leveldb::in_memory()).unwrap(),
-                ))
+            .map(|_| {
+                RustyLevelDbShim::new(
+                    ":memory",
+                )
             })
             .collect()
     }
 
-    fn create_core_dbs(dbs: &Vec<Arc<Mutex<rusty_leveldb::DB>>>) -> Vec<Arc<RwLock<CoreEthereumDb<RustyLevelDbShim>>>> {
+    fn create_core_dbs(dbs: &Vec<RustyLevelDbShim>) -> Vec<Arc<RwLock<CoreEthereumDb<RustyLevelDbShim>>>> {
         dbs.iter()
             .enumerate()
             .map(|(i, db)| {
                 Arc::new(RwLock::new(CoreEthereumDb::new(
-                    DB::new(RustyLevelDbShim::new(db.clone())),
+                    DB::new(db.clone()),
                     PublicKey::from_privkey(&PEERS_CHAIN_PRIVS[i]).unwrap().to_address(),
                 )))
             })
             .collect::<Vec<_>>()
     }
 
-    async fn create_minimal_topology(dbs: &Vec<Arc<Mutex<rusty_leveldb::DB>>>) -> crate::errors::Result<()> {
+    async fn create_minimal_topology(dbs: &Vec<RustyLevelDbShim>) -> crate::errors::Result<()> {
         let testing_snapshot = Snapshot::default();
         let mut previous_channel: Option<ChannelEntry> = None;
 
         for index in 0..dbs.len() {
             let mut db = CoreEthereumDb::new(
-                DB::new(RustyLevelDbShim::new(dbs[index].clone())),
+                DB::new(dbs[index].clone()),
                 PEERS_CHAIN[index].public().to_address(),
             );
