+++ packages/core/crates/core-packet/src/validation.rs
@@ -118,7 +118,7 @@ mod tests {
     use mockall::mock;
     use std::sync::{Arc, Mutex};
     use utils_db::db::DB;
-    use utils_db::leveldb::rusty::RustyLevelDbShim;
+    use utils_db::rusty::RustyLevelDbShim;
     use utils_types::primitives::{
         Address, AuthorizationToken, Balance, BalanceType, EthereumChallenge, Snapshot, U256,
     };
@@ -570,11 +570,8 @@ mod tests {
 
     #[async_std::test]
     async fn test_ticket_workflow() {
-        let level_db = Arc::new(Mutex::new(
-            rusty_leveldb::DB::open("test", rusty_leveldb::in_memory()).unwrap(),
-        ));
         let mut db = CoreEthereumDb::new(
-            DB::new(RustyLevelDbShim::new(level_db)),
+            DB::new(RustyLevelDbShim::new(":memory")),
             SENDER_PRIV_KEY.public().to_address(),
         );
 
@@ -616,11 +613,8 @@ mod tests {
 
     #[async_std::test]
     async fn test_db_should_store_ticket_index() {
-        let level_db = Arc::new(Mutex::new(
-            rusty_leveldb::DB::open("test", rusty_leveldb::in_memory()).unwrap(),
-        ));
         let mut db = CoreEthereumDb::new(
-            DB::new(RustyLevelDbShim::new(level_db)),
+            DB::new(RustyLevelDbShim::new(":memory")),
             SENDER_PRIV_KEY.public().to_address(),
         );
