+++ packages/hoprd/src/api/v3/paths/token.integration.spec.ts
@@ -3,7 +3,7 @@ import sinon from 'sinon'
 import chaiResponseValidator from 'chai-openapi-response-validator'
 import chai, { expect } from 'chai'
 
-import { LevelDb, Database } from '@hoprnet/hopr-utils'
+import { Database } from '@hoprnet/hopr-utils'
 
 import { createToken, storeToken } from '../../token.js'
 import { createAuthenticatedTestApiInstance, ALICE_ETHEREUM_ADDR } from './../fixtures.js'
@@ -16,9 +16,9 @@ describe('GET /token', function () {
 
   before(async function () {
     node = sinon.fake() as any
-    let db = new LevelDb()
-    await db.backend.open()
-    node.db = new Database(db, ALICE_ETHEREUM_ADDR.clone())
+    /*let db = new LevelDb()
+    await db.backend.open()*/
+    node.db = new Database(":memory", ALICE_ETHEREUM_ADDR.clone())
 
     const loaded = await createAuthenticatedTestApiInstance(node)
