+++ packages/utils/crates/utils-db/src/lib.rs
@@ -5,6 +5,7 @@ pub mod hashmap;
 pub mod leveldb;
 pub mod traits;
 pub mod types;
+pub mod rusty;
 
 pub use traits::KVStorage;
